requirements.txt
RPi.GPIO
opencv-python
opencv-contrib-python
numpy
